BACKUP_BUG18_FIXED_2025-12-17_18-02-25/requirements.txt
aiogram==3.13.1
python-dotenv==1.0.0
aiosqlite==0.19.0
